+++ src/main/java/com/copypay/listener/CustomAuthFailureListener.java
@@ -1,4 +1,4 @@
-package com.copypay.config;
+package com.copypay.listener;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationListener;
