+++ src/main/java/com/copypay/listener/CustomAuthSuccessListener.java
@@ -1,4 +1,4 @@
-package com.copypay.config;
+package com.copypay.listener;
 
 import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
